Moving Pagination class to its own file
